이미지 route 생성
