+++ backend/controllers/user.controller.js
@@ -187,28 +187,4 @@ function uploadImage(req, res, next) {
   });
 }
 
-function profileImage(req, res, next) {
-  fs.readFile(profileImagePath + "profile_" + req.params.userId + ".jpg", function(error,data){
-    if (error) {
-      res.writeHead(404, {"Content": "image/jpeg"})
-      res.end()
-    } else {
-      res.writeHead(200, {"Content": "image/jpeg"})
-      res.end(data)
-    }
-  })
-}
-
-function profileThumbnail(req, res, next) {
-  fs.readFile(profileThumbnailImagePath + "profile_" + req.params.userId + "-50x50.jpg", function(error,data){
-    if (error) {
-      res.writeHead(404, {"Content": "image/jpeg"})
-      res.end()
-    } else {
-      res.writeHead(200, {"Content": "image/jpeg"})
-      res.end(data)
-    }
-  })
-}
-
-module.exports = { load, get, create, update, list, remove, getTokens, getMyToken, activeList, addressList, uploadImage, profileImage, profileThumbnail };
+module.exports = { load, get, create, update, list, remove, getTokens, getMyToken, activeList, addressList, uploadImage };
