+++ backend/routes/user.route.js
@@ -25,11 +25,7 @@ router.route('/address')
   .get(auth, userCtrl.addressList)
 
 router.route('/:userId/images/profile')
-  .get(userCtrl.profileImage)
   .post(auth, userCtrl.uploadImage)
-
-router.route('/:userId/images/profile/thumbnail')
-  .get(userCtrl.profileThumbnail)
     
 router.route('/me')
   /** GET /api/users/me - Get current user */
